requirements.txt
python-telegram-bot==5.2.0
requests==2.12.0
